refactor(install): extract setup core helpers
