+++ package.json
@@ -29,10 +29,11 @@
     "smoke": "npm run smoke:analytics",
     "smoke:analytics": "npm run build && node scripts/smoke-analytics.mjs",
     "test": "npm run test:all",
-    "test:all": "npm run test:worker && npm run test:registry && npm run test:install && npm run test:maintenance && npm run test:upstream-release && npm run test:build-core && npm run test:build-html-core",
+    "test:all": "npm run test:worker && npm run test:registry && npm run test:install && npm run test:install-core && npm run test:maintenance && npm run test:upstream-release && npm run test:build-core && npm run test:build-html-core",
     "test:build-core": "node scripts/build-site-core.test.mjs",
     "test:build-html-core": "node scripts/build-html-core.test.mjs",
     "test:install": "node scripts/install.test.mjs",
+    "test:install-core": "node scripts/install-core.test.mjs",
     "test:maintenance": "node scripts/maintenance.test.mjs",
     "test:registry": "node scripts/registry.test.mjs",
     "test:upstream-release": "node scripts/upstream-release.test.mjs",
